Implemented ccn-lite conent format according to ccn-lite-content-format.md and some fix for varNum in NDNTLV

+++ src/main/scala/ccn/ccnlite/ndntlv/tlvtranscoding/TLV.scala
@@ -1,3 +1,3 @@
 package ccn.ccnlite.ndntlv.tlvtranscoding
 
-case class TLV(t: Int, l: Int, v: List[Byte])
+case class TLV(t: Long, l: Long, v: List[Byte])
